+++ apps/web/app/(main)/todos/_components/create-todo-form/create-todo-controller.test.ts
@@ -1,5 +1,5 @@
 import { type Mock, describe, expect, it, vi } from "vitest";
-import { createTodoController } from "./controller";
+import { createTodoController } from "./create-todo-controller";
 
 type PresenterMock = {
   presentFormData: Mock;
