+++ apps/web/app/(main)/todos/_components/create-todo-form/create-todo-form.tsx
@@ -4,7 +4,7 @@ import { Box, Button, Flex, TextArea, TextField } from "@radix-ui/themes";
 import { type JSX, useActionState } from "react";
 import { ErrorMessagePanel } from "#components/message-panel";
 import { SessionUserId } from "#lib/session-types";
-import { createTodo } from "./action";
+import { createTodo } from "./create-todo-action";
 
 type Props = {
   readonly userId: SessionUserId;
